float64 quicker?

+++ Assignment1/code/viterbi.py
@@ -3,7 +3,7 @@ import utils
 
 def run_viterbi_2nd_order_log_with_beam_search(sentence_words, words_count, tags_count, start_tag_id, getLogScore):
     """Note: tags_count includes START TAG"""
-    V = np.full([words_count + 1, tags_count, tags_count], float('-inf'), dtype=np.float32)
+    V = np.full([words_count + 1, tags_count, tags_count], float('-inf'), dtype=np.float64)
     bp = np.full([words_count, tags_count, tags_count], -1, dtype=np.int32)
 
     V[0, start_tag_id, start_tag_id] = np.log(1)
